requirements.txt
click==6.6
Flask==0.11.1
futures==3.0.5
gnupg==2.0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
psutil==4.3.0
pymongo==3.2.2
python-swiftclient==3.0.0
requests==2.10.0
six==1.10.0
Werkzeug==0.11.10
wheel==0.24.0
python-keystoneclient==3.9.0
pycrypto==2.6.1
cloudant==2.3.1